Move MockAction and ActionWithPayload types to src/types/action.tsx
