+++ src/components/GoalView/tests/MockAction.tsx
@@ -1,5 +1,5 @@
 import { Goal } from "../../../types/goals";
-import { ActionWithPayload } from "../GoalTimelineActions";
+import { ActionWithPayload } from "../../../types/action";
 
 export const MOCK_GOAL = "MOCK_GOAL";
 export type MOCK_GOAL = typeof MOCK_GOAL;
