+++ src/components/Nav/navActions.tsx
@@ -1,14 +1,9 @@
 import { Dispatch } from "react";
-import { Action } from "redux";
 import { Goal } from "../../types/goals";
+import { ActionWithPayload } from "../../types/action";
 
 export const CHANGE_DISPLAY = "CHANGE_DISPLAY";
 export type CHANGE_DISPLAY = typeof CHANGE_DISPLAY;
-
-// action types
-export interface ActionWithPayload<T> extends Action {
-  payload: T;
-}
 
 export interface ChangeDisplay extends ActionWithPayload<Goal> {
   type: CHANGE_DISPLAY;
